nami/requirements.txt
asgiref==3.2.10
pytz==2020.1
dj-database-url==0.5.0
Django==3.0
django-heroku==0.3.1
gunicorn==20.0.4
python-decouple==3.3
sqlparse==0.3.1
whitenoise==5.2.0
pipenv==2020.11.15
djangorestframework
django-cors-headers 

